# Improve siphash performance for small keys by up to 30%